requirements.txt
aiofiles
aiohttp
asyncio
dnspython
ffmpeg-python
gitpython
hachoir
heroku3
lyricsgenius
motor==3.0.0
pillow
psutil
py-tgcalls==0.8.6
pykeyboard
pyrogram==1.4.16
python-dotenv
pyyaml==5.4
requests
speedtest-cli
spotipy
tgcrypto
numpy
youtube-search
youtube-search-python
git+https://github.com/yt-dlp/yt-dlp
